+++ fts_examples/test_examples.py
@@ -46,6 +46,7 @@ if Version(torch_version) < Version(MAX_VERSION_WARNS):
     EXPECTED_WARNS.extend(EXPECTED_VERSION_WARNS[MIN_VERSION_WARNS])
 else:
     EXPECTED_WARNS.extend(EXPECTED_VERSION_WARNS[MAX_VERSION_WARNS])
+ADV_EXPECTED_WARNS = EXPECTED_WARNS + ["Found an `init_pg_lrs` key"]
 
 
 @pytest.mark.skipif(not _HF_AVAILABLE, reason="Hugging Face transformers and datasets packages required")
@@ -70,4 +71,29 @@ def test_examples_fts_superglue(monkeypatch, recwarn, tmpdir, config_file):
     cli_main()
     # ensure no unexpected warnings detected
     matched = [any([re.compile(w).search(w_msg.message.args[0]) for w in EXPECTED_WARNS]) for w_msg in recwarn.list]
+    assert all(matched)
+
+
+@pytest.mark.skipif(not _HF_AVAILABLE, reason="Hugging Face transformers and datasets packages required")
+@RunIf(min_gpus=1, skip_windows=True)
+@pytest.mark.parametrize(
+    "config_file",
+    [pytest.param("fts_explicit_reinit_lr.yaml", marks=RunIf(min_torch="1.10")), "fts_implicit_reinit_lr.yaml"],
+    ids=["fts_explicit_reinit_lr", "fts_implicit_reinit_lr"],
+)
+def test_advanced_examples_fts_superglue(monkeypatch, recwarn, tmpdir, config_file):
+    from fts_examples.fts_superglue import cli_main
+
+    example_script = os.path.join(os.path.dirname(__file__), "fts_superglue.py")
+    config_loc = [os.path.join(os.path.dirname(__file__), "config/advanced", config_file)]
+    cli_args = [
+        f"--trainer.default_root_dir={tmpdir.strpath}",
+        "--trainer.max_epochs=10",
+        "--trainer.limit_train_batches=2",
+        "--trainer.gpus=1",
+    ]
+    monkeypatch.setattr("sys.argv", [example_script, "fit", "--config"] + config_loc + cli_args)
+    cli_main()
+    # ensure no unexpected warnings detected
+    matched = [any([re.compile(w).search(w_msg.message.args[0]) for w in ADV_EXPECTED_WARNS]) for w_msg in recwarn.list]
     assert all(matched)